Trait smtp_server::Config

pub trait Config: Send + Sync {
    type ConnectionUserMeta: Send;
    type MailUserMeta: Send;

    fn hostname(
        &self, 
        conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> &str;

    #[must_use]
    fn tls_accept<'life0, 'life1, 'async_trait, IO>(
        &'life0 self, 
        io: IO, 
        conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Result<Duplex<Pin<Box<dyn Send + AsyncRead>>, Pin<Box<dyn Send + AsyncWrite>>>>> + Send + 'async_trait>>
    where
        IO: 'static + Unpin + Send + AsyncRead + AsyncWrite,
        IO: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn new_mail<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Self::MailUserMeta> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn filter_from<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        from: Option<Email>, 
        meta: &'life1 mut MailMetadata<Self::MailUserMeta>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<Option<Email>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn filter_to<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        to: Email, 
        meta: &'life1 mut MailMetadata<Self::MailUserMeta>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<Email>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn handle_mail<'a, 'life0, 'life1, 'life2, 'async_trait, R>(
        &'life0 self, 
        stream: &'life1 mut EscapedDataReader<'a, R>, 
        meta: MailMetadata<Self::MailUserMeta>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        R: Send + Unpin + AsyncRead,
        'a: 'async_trait,
        R: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn welcome_banner(
        &self, 
        conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> &str { ... }

    fn welcome_banner_reply(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn hello_banner(
        &self, 
        conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> &str { ... }

    #[must_use]
    fn filter_hello<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        is_ehlo: bool, 
        hostname: Hostname, 
        conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<HelloInfo>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn can_do_tls(
        &self, 
        conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> bool { ... }

    #[must_use]
    fn filter_data<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        meta: &'life1 mut MailMetadata<Self::MailUserMeta>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    fn handle_rset<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        meta: &'life1 mut Option<MailMetadata<Self::MailUserMeta>>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    fn handle_starttls<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    fn handle_expn<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        name: MaybeUtf8<&'life1 str>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    fn handle_vrfy<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        name: MaybeUtf8<&'life1 str>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    fn handle_help<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        subject: MaybeUtf8<&'life1 str>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    fn handle_noop<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        string: MaybeUtf8<&'life1 str>, 
        conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    fn handle_quit<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn already_did_hello(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn mail_before_hello(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn already_in_mail(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn rcpt_before_mail(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn data_before_rcpt(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn data_before_mail(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn starttls_unsupported(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn command_unrecognized(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn pipeline_forbidden_after_starttls(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn line_too_long(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn handle_mail_did_not_call_complete(
        &self, 
        conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
    ) -> Reply { ... }

    fn reply_write_timeout(&self) -> Duration { ... }

    fn command_read_timeout(&self) -> Duration { ... }

}

Associated Types

type ConnectionUserMeta: Send

type MailUserMeta: Send

Loading content...
Required methods

fn hostname(
&self,
conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
) -> &str

Note: this function is only ever used for the default implementations of other functions in this trait. As such, it is OK to leave it unimplemented!() if other functions are implemented.

#[must_use]
fn tls_accept<'life0, 'life1, 'async_trait, IO>(
&'life0 self,
io: IO,
conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Result<Duplex<Pin<Box<dyn Send + AsyncRead>>, Pin<Box<dyn Send + AsyncWrite>>>>> + Send + 'async_trait>> where
IO: 'static + Unpin + Send + AsyncRead + AsyncWrite,
IO: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Note: if you don’t want to implement TLS, you should override can_do_tls to return false so that STARTTLS is not advertized. This being said, returning an error here should have the same result in practice, except clients will try STARTTLS and fail

#[must_use]
fn new_mail<'life0, 'life1, 'async_trait>(
&'life0 self,
conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Self::MailUserMeta> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

#[must_use]
fn filter_from<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
from: Option<Email>,
meta: &'life1 mut MailMetadata<Self::MailUserMeta>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<Option<Email>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn filter_to<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
to: Email,
meta: &'life1 mut MailMetadata<Self::MailUserMeta>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<Email>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_mail<'a, 'life0, 'life1, 'life2, 'async_trait, R>(
&'life0 self,
stream: &'life1 mut EscapedDataReader<'a, R>,
meta: MailMetadata<Self::MailUserMeta>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
R: Send + Unpin + AsyncRead,
'a: 'async_trait,
R: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Note: the EscapedDataReader has an inner buffer size of RDBUF_SIZE, which means that reads should not happen with more than this buffer size.

Also, note that there is no timeout applied here, so the implementation of this function is responsible for making sure that the client does not just stop sending anything to DOS the system.

Loading content...
Provided methods

fn welcome_banner(
&self,
conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
) -> &str

Note: this function is only ever used for the default implementations of other functions in this trait. As such, it is OK to leave it unimplemented!() if other functions are implemented.

fn welcome_banner_reply(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn hello_banner(
&self,
conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
) -> &str

Note: this function is only ever used for the default implementations of other functions in this trait. As such, it is OK to leave it unimplemented!() if other functions are implemented.

#[must_use]
fn filter_hello<'life0, 'life1, 'async_trait>(
&'life0 self,
is_ehlo: bool,
hostname: Hostname,
conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<HelloInfo>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn can_do_tls(
&self,
conn_meta: &ConnectionMetadata<Self::ConnectionUserMeta>
) -> bool

#[must_use]
fn filter_data<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
meta: &'life1 mut MailMetadata<Self::MailUserMeta>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_rset<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
meta: &'life1 mut Option<MailMetadata<Self::MailUserMeta>>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_starttls<'life0, 'life1, 'async_trait>(
&'life0 self,
conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_expn<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
name: MaybeUtf8<&'life1 str>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_vrfy<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
name: MaybeUtf8<&'life1 str>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_help<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
subject: MaybeUtf8<&'life1 str>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_noop<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
string: MaybeUtf8<&'life1 str>,
conn_meta: &'life2 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]
fn handle_quit<'life0, 'life1, 'async_trait>(
&'life0 self,
conn_meta: &'life1 mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Pin<Box<dyn Future<Output = Decision<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn already_did_hello(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn mail_before_hello(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn already_in_mail(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn rcpt_before_mail(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn data_before_rcpt(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn data_before_mail(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn starttls_unsupported(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn command_unrecognized(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn pipeline_forbidden_after_starttls(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn line_too_long(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn handle_mail_did_not_call_complete(
&self,
conn_meta: &mut ConnectionMetadata<Self::ConnectionUserMeta>
) -> Reply

fn reply_write_timeout(&self) -> Duration

fn command_read_timeout(&self) -> Duration

Loading content...