Module smtp_server::reply

Functions

bad_sequence

Usual value for returning “Okay” from already_did_hello, mail_before_hello, already_in_mail, rcpt_before_mail, data_before_rcpt and data_before_mail

command_not_supported

command_unimplemented

command_unrecognized

handle_mail_did_not_call_complete

ignore_help

Usual value for ignoring the request but returning a generic message from handle_help

ignore_vrfy

Usual value for ignoring the request but returning “Okay” from handle_vrfy

internal_server_error

line_too_long

okay

okay_data

Usual value for returning “Okay” from filter_data

okay_from

Usual value for returning “Okay” from filter_from

okay_hello

Usual value for returning “Okay” from filter_hello

okay_mail

Usual value for returning “Okay” from handle_mail

okay_noop

Usual value for returning “Okay” from handle_noop

okay_quit

Usual value for returning “Okay” from handle_quit

okay_rset

Usual value for returning “Okay” from handle_rset

okay_starttls

Usual value for returning “Okay” from handle_starttls

okay_to

Usual value for returning “Okay” from filter_to

pipeline_forbidden_after_starttls

welcome_banner